refactor: restructure project by vertical slicing
