+++ core/src/main/java/com/dam/commune/community/CommunityRepository.java
@@ -1,9 +1,7 @@
-package com.dam.commune.repository;
+package com.dam.commune.community;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.dam.commune.entity.Community;
 
 @Repository
 public interface CommunityRepository extends JpaRepository<Community, Long> {
