+++ core/src/main/java/com/dam/commune/community/CommunityService.java
@@ -1,9 +1,7 @@
-package com.dam.commune.service;
+package com.dam.commune.community;
 
 import java.util.List;
 import java.util.Optional;
-
-import com.dam.commune.entity.Community;
 
 public interface CommunityService {
     List<Community> findAll();
